refactor: finish table component
