+++ src/components/features/ui-elements/examples/index.ts
@@ -2,18 +2,18 @@ import { Chat } from "./chat";
 import { FinancialPerformanceChart } from "./financial-performance-chart";
 import { MusicPlayer } from "./music-player";
 import { NotificationSettings } from "./notification-settings";
-import DataTableDemo from "./payment-management/example-2";
+import DataTableDemo from "./payment-management";
 import { QuarterlyFinancialReviewChart } from "./quarterly-financial-review";
 import { ShareDocument } from "./share-document";
 import { UserSelect } from "./user-select";
 
 export const UIElementsExamples = [
-    // MusicPlayer,
-    // UserSelect,
-    // QuarterlyFinancialReviewChart,
-    // FinancialPerformanceChart,
-    // Chat,
-    // NotificationSettings,
-    // ShareDocument,
-    DataTableDemo
+  MusicPlayer,
+  UserSelect,
+  QuarterlyFinancialReviewChart,
+  FinancialPerformanceChart,
+  Chat,
+  NotificationSettings,
+  ShareDocument,
+  DataTableDemo,
 ];
