+++ src/components/features/ui-elements/examples/payment-management/constants/index.ts
@@ -1,9 +1,9 @@
 import { ChipProps } from "@nextui-org/react";
 
-export const INITIAL_VISIBLE_COLUMNS = ["name", "role", "status", "actions"];
+export const INITIAL_VISIBLE_COLUMNS = ["name"];
 
-export const STATUS_COLOR_MAP: Record<string, ChipProps["color"]>  = {
-    active: "success",
-    paused: "danger",
-    vacation: "warning",
-  };
+export const STATUS_COLOR_MAP: Record<string, ChipProps["color"]> = {
+  active: "success",
+  paused: "danger",
+  vacation: "warning",
+};
